+++ backend/bookmarks/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from users.models import CustomUser
-from users.serializers import CustomUserTelegramIDSerializer, CustomUserSerializer
+from users.serializers import CustomUserSerializer, CustomUserTelegramIDSerializer
 
 from .models import Bookmark, Folder, Tag
 
@@ -172,13 +172,39 @@ class BookmarkUpdateSerializer(serializers.ModelSerializer):
         """
         Override `update` to deal with nested objects.
         """
-        instance.url = validated_data.get("url")
-        instance.title = validated_data.get("title")
-        instance.description = validated_data.get("description")
-        instance.image_url = validated_data.get("image_url")
-        instance.is_favorite = validated_data.get("is_favorite")
-        instance.is_read = validated_data.get("is_read")
-        instance.is_archived = validated_data.get("is_archived")
+        instance.url = (
+            validated_data.get("url") if validated_data.get("url") else instance.url
+        )
+        instance.title = (
+            validated_data.get("title")
+            if validated_data.get("title")
+            else instance.title
+        )
+        instance.description = (
+            validated_data.get("description")
+            if validated_data.get("description")
+            else instance.description
+        )
+        instance.image_url = (
+            validated_data.get("image_url")
+            if validated_data.get("image_url")
+            else instance.image_url
+        )
+        instance.is_favorite = (
+            validated_data.get("is_favorite")
+            if validated_data.get("is_favorite")
+            else instance.is_favorite
+        )
+        instance.is_read = (
+            validated_data.get("is_read")
+            if validated_data.get("is_read")
+            else instance.is_read
+        )
+        instance.is_archived = (
+            validated_data.get("is_archived")
+            if validated_data.get("is_archived")
+            else instance.is_archived
+        )
 
         # Take care of nested objects
         if folder_data := validated_data.get("folder"):